\name{closeViz-methods}
\docType{methods}
\alias{closeViz-methods}
\alias{closeViz,NG_Visualization-method}
\alias{closeViz,NG_Viz2DAxis-method}
\alias{closeViz,NG_Viztk2d-method}

\title{Initialization of a new Display}
\description{
If custom visualization class (NG_Visualization) has been defined. This is the method that gets called if navGraph switches away from a graph that links to the visualization class.

See package vignette for more detail.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "NG_Visualization")}}{}

}}
\keyword{methods}
