\name{initializeViz}
\alias{initializeViz}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Method that gets called for initializing a display for some visualization instructions
}
\description{
   See the vignette for more details.
}
\usage{
initializeViz(viz,ngEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{viz}{
	Object of class NG_Visualization
}
  \item{ngEnv}{
   Environment of navGraph session.
}
}
%\details{ }
\value{
  Object of NG_Visualization
}

\author{
Adrian Waddell and Wayne Oldford
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{updateViz}}, \code{\link{closeViz}}
}
\examples{
## See vignette
}

