\name{linegraph-methods}
\docType{methods}
\alias{linegraph-methods}
\alias{linegraph,graph-method}
\title{ Linegraph of a graph }
\description{
  (From Wikipedia) The line graph L(G) of an undirected graph G is another graph L(G) that represents the adjacencies between edges of G. By definition, each vertex of L(G) represents an edge of G, and two vertices of L(G) are adjacent if and only if their corresponding edges share a common endpoint ("are adjacent") in G.
}
\section{Methods}{
\describe{
  \item{\code{signature(x = "NG_data")}}{  }
}}
\keyword{methods}