\name{names-methods}
\docType{methods}
\alias{names-methods}
\alias{names,NG_data-method}
\title{ Show the variable names of a NG_data object}
\description{
  Show the variable names of a NG_data object. 
}
\section{Methods}{
\describe{

\item{\code{signature(x = "NG_data")}}{  }
}}
\keyword{methods}

