\name{names<--methods}
\docType{methods}
%\alias{names<--methods}
\alias{names<-,NG_data,ANY-method}
\title{Change the names of a NG_data object}
\description{
  Change the names of a NG_data object. This end effectively changes the names of the data.frame wrapped in the NG_data object.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "NG_data")}}{  }
}}
\keyword{methods}
