\name{ng_get-methods}
\docType{methods}
\alias{ng_get-methods}
\alias{ng_get,NavGraph_handler-method}
\alias{ng_get,NG_data-method}
\alias{ng_get,NG_graph-method}
\alias{ng_get,NG_path-method}
\title{Access data from a NG_... object}
\description{
  Access data in the wrapper classes in navGraph or the navGraph
  hanlder.

  The usage is either ng_get(object) or ng_get(object,"what").

  See in the examples of \code{\link{ng_update}}.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "NavGraph_handler")}}{}

\item{\code{signature(obj = "NG_data")}}{}

\item{\code{signature(obj = "NG_graph")}}{}

\item{\code{signature(obj = "NG_path")}}{}
}}
\keyword{methods}
