\name{ng_set_color<--methods}
\docType{methods}
\alias{ng_set_color<--methods}
\alias{ng_set_color<-,NavGraph_handler-method}
\title{ Change colors of data points in an active navGraph session}
\description{
	Specify new colors for each point for an active navGraph session.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "NavGraph_handler")}}{ }

}}
\keyword{methods}
