\name{ng_set_size<--methods}
\docType{methods}
\alias{ng_set_size<--methods}
\alias{ng_set_size<-,NavGraph_handler-method}
\title{ Change sizes of data points in an active navGraph session}
\description{
	Specify new sizes for each point for an active navGraph session.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "NavGraph_handler")}}{ }

}}
\keyword{methods}
