\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,NG_graph-method}
\title{ Plot the graph saved in a NG_data object.}
\description{
	Once the graph is within a NG_graph object, it has a layout associated. Plot plots the graph with this layout.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "NG_data")}}{ }
}}
\keyword{methods}
