\name{shortnames-methods}
\docType{methods}
\alias{shortnames-methods}
\alias{shortnames,NG_data-method}
\title{ Show the shortnames of a NG_data object}
\description{
  The variable names coded in the node names of a graph in a
  \code{navGraph} session will be matched either with the names or
  shortnames of a NG_data object. 
}
\section{Methods}{
\describe{
  \item{\code{signature(x = "NG_data")}}{  }
}}
\keyword{methods}

