\name{shortnames<--methods}
\docType{methods}
\alias{shortnames<--methods}
\alias{shortnames<-,NG_data-method}
\title{Change the shortnames of a NG_data object}
\description{
  Change the shortnames of a NG_data object.

  Note that the shortnames of a NG_data object can not contain any spaces.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "NG_data")}}{  }
}}
\keyword{methods}
