\name{updateViz-methods}
\docType{methods}
\alias{updateViz-methods}
\alias{updateViz,NG_Visualization-method}
\alias{updateViz,NG_Viz2DAxis-method}
\alias{updateViz,NG_Viztk2d-method}

\title{Initialization of a new Display}
\description{
If custom visualization class (NG_Visualization) has been defined. This is the method that gets called if the bullet changes its state.

See package vignette for more detail.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "NG_Visualization")}}{}

}}
\keyword{methods}
